Rename styles directory to style-definitions
